Hacking on Babel mocking:
Status quo
a BufStream-backed Babel struct will work in testing
but not in integration-tests.

Old behavior
The motivation for moving to BufStream was that once an ad-hoc
instantiated BufReader would go out of scope in babel_read, so would the
prefetched bytes of the buffer - i.e. if you preemptively pushed two
responses to a SharedMockStream and tried doing two `command`s or
`read_babel`s on it, it would take the whole buffer, load it into a
BufReader and consume it until the first "ok". Notice that the other
response with the second "ok" is already in the BufReader here.
Then the BufReader would go out of scope, while the internal stream's
output would already be exhausted into the BufReader's internal
buffering. Effectively, real babel would work cause there'd be no more
than one command output to read per command, while the tests wouldn't in
all the cases where you push two command outputs.

New behavior
Instead of being a trait, Babel is now a struct that stores a BufStream
interanally thus (in theory) preserving the buffering state between
reads.

What's the problem?
When coing rita.sh Rita reads Babel's preamble no problem, but when it
sends `interface wg0` to the socket, it can't read the response. The
weirdest thing is that a SharedMockStream works no problem.